Fix compilation warnings of examples

+++ examples/digit/main.c
@@ -1,7 +1,7 @@
 #include <Arduino.h>
 
-byte digits []={0xFC,0x60,0xDA,0xF2,0x66,0xB6,0xBE,0xE0,0xFE,0xF6};
-int  led2pin[]={9, 8, 7,10,11,12,13, 6};
+uint8_t digits []={0xFC,0x60,0xDA,0xF2,0x66,0xB6,0xBE,0xE0,0xFE,0xF6};
+uint8_t  led2pin[]={9, 8, 7,10,11,12,13, 6};
 void displayDigit(int d)
 {
 	byte code = digits[d%10];
@@ -14,7 +14,7 @@ void displayDigit(int d)
 
 void setup()
 {
-	for (int i = 6; i <= 13; i++)
+	for (uint8_t i = 6; i <= 13; i++)
 		pinMode(i, OUTPUT);
 }
 void loop()
